requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
ultralytics>=8.0.0
tqdm>=4.65.0
nltk>=3.8.0
